CP-307 fix: fix favorite page
