+++ src/api/asktug/profile.ts
@@ -16,6 +16,53 @@ export interface IRawBadges {
   has_badge: boolean;
   long_description: string;
 }
+export interface IBlogAuthor {
+  id: number;
+  username: string;
+  avatarURL: string;
+}
+
+//export interface IBlogCategory {
+//  id: number;
+//  name: string;
+//  slug: string;
+//}
+
+export interface IPost {
+  id: number;
+  slug: string;
+  status: string;
+  author: IBlogAuthor;
+  origin: string;
+  title: string;
+  summary: string;
+  publishedAt?: Date;
+  createdAt: Date;
+  lastModifiedAt: Date;
+  deletedAt?: Date;
+  recommended: boolean;
+  recommendAt?: Date;
+  tags: [];
+  likes: number;
+  comments: number;
+}
+
+export interface IPostFavorite {
+  post: IPost;
+  user: IBlogAuthor;
+}
+
+export interface IPage {
+  number: number;
+  size: number;
+  totalElements: number;
+  totalPages: number;
+}
+
+export interface IResponse<T> {
+  content: T[];
+  page: IPage;
+}
 
 async function getAllBadges(): Promise<Map<IRawBadges['id'], IRawBadges>> {
   const result: { badges: any[] } = await asktugClient.get(`${askTugApiDomain}/badges.json`);
